Manage cross-file generics
